+++ packages/yourloops/components/routes.tsx
@@ -34,10 +34,11 @@ import { ThemeProvider } from "@material-ui/core/styles";
 
 import { UserRoles } from "../models/shoreline";
 import { useAuth } from "../lib/auth";
+import { setPageTitle } from "../lib/utils";
+import { NotificationContextProvider } from "../lib/notifications/hook";
+import { SnackbarContextProvider, DefaultSnackbarContext } from "./utils/snackbar";
 import { externalTheme, mainTheme } from "./theme";
 import FooterLinks from "./footer-links";
-import { SnackbarContextProvider, DefaultSnackbarContext } from "./utils/snackbar";
-import { NotificationContextProvider } from "../lib/notifications/hook";
 
 export const PublicRoute = (props: RouteProps): JSX.Element => {
   const historyHook = useHistory<{ from?: { pathname?: string; }; }>();
@@ -46,6 +47,12 @@ export const PublicRoute = (props: RouteProps): JSX.Element => {
     const fromPath = historyHook.location.state?.from?.pathname;
     return <Redirect to={{ pathname: fromPath ?? user?.getHomePage() }} />;
   }
+
+  // eslint-disable-next-line react-hooks/rules-of-hooks
+  React.useEffect(() => {
+    setPageTitle();
+  });
+
   return (
     <ThemeProvider theme={externalTheme}>
       <CssBaseline />