+++ packages/yourloops/pages/login/login.tsx
@@ -116,10 +116,6 @@ function Login(): JSX.Element {
     }
   }, [signupEmail, username]);
 
-  React.useEffect(() => {
-    document.title = t("brand-name");
-  }, [t]);
-
   const onClickLoginButton = async (): Promise<void> => {
     if (_.isEmpty(username) || _.isEmpty(password)) {
       setValidateError(true);